+++ CHANGELOG.md
@@ -9,6 +9,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- support System Suspend extension
+- support CPPC extension 
+
 ### Modified
 
 - run on provided `MachineInfo` by default; bare metal M-mode environment should gate `machine`